requirements.txt
fastapi==0.99.1
uvicorn==0.23.2
python-dotenv==1.0.1

langchain==0.1.20
langchain-core==0.1.53
langchain-community==0.0.38
langchain-groq==0.1.5
langchain-huggingface==0.0.3

sentence-transformers==2.6.1
transformers==4.41.2
huggingface-hub==0.36.2

chromadb==0.4.24
pypdf==3.17.4
beautifulsoup4==4.12.3
requests==2.31.0
